[김시아] #Feat : 나의 발자취 조회 API 완료

+++ back/ollana/src/main/java/com/ssafy/ollana/footprint/web/dto/response/FootprintResponseDto.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @Getter
 @Builder
 public class FootprintResponseDto {
-    private Long footprintId;
+    private Integer footprintId;
     private String mountainName;
     private String imgUrl;
 }
